Adds Landing package.
Adds `Lander.{cpp,h}` and `Parachute.{cpp,h}` to represent the Landing sub-package within the Landing package. In this commit, the Lander.c repersents the entire Landing package (handling all events handling related) whereas the Parachute class acts as a complex way to interface with the parachute of the rocket. A Parachute is stored as a private field within the Lander class. As such, this should be the only way that the parachute is interfaced with.