Adds Gimbal package.
Adds `Gimbal.{cpp,h}` and `Servo.{cpp,h}` to represent the Gimabl sub-package within the Control package on the software achitecture diagram. In this commit, the Gimbal.h describes the functionality of the gimbal of the rocket entirely where Servo.h is soley a complex representation of the servos within the Gimbal itseft. This relationship is very similar to what is identified within the Landing package where Servo objects are stored privately as fields within the Gimbal class and this is the only way that they should be interfaced with.