Adds pseudo-code implementation of methods in control package
In this commit we implement some pseudo-code methods for the control package.
Currently these implementations log to the console what the software is doing.
Where appropriate, these methods perform as they may down the track, however,
a lot of the methods still contain arbitrary code and comments with the intent
of being implemeneted down the track. To accompany this, throughout the files
some TODO(ALL): tags have been added to indicate work is still needed here.