Fixes C++ issues and adds a test file
In this commit we fix any C++ violations ensuring all headers
and files and compile using G++. In this commit, we also add a
tests.cpp file. This creates objects from all of the classes
declared within the package. This also runs their methods to
ensure that there are no errors present. In the future, we will
focus on researching into unit testing frameworks for C++. To
help build, lint and compile the tests a bash script run_tests.sh
has been added. This requires the user has cpplint and gcc
installed on the system