Refactors software package to comply with Platformio.
In this commit, we add a working main cpp file that runs on the board (Credit to
Niels on getting this working). We shuffle around the existing files to better comply
with the recommended directory structures. We also add an entry the the .gitlab-ci.yml
which should hopefully be on the right track (but I suspect that there will be issues
with this). We also add an example test file that follows the recommended structure and
includes files from the control library. This is a work in progress, and a lot of the
stuff is very finicky.