Adds platformio main executable.
In this commit we add a working main.cpp. This executable configures most of the
hardware as well as prints out the x axis acceleration of the MPU. This file
contains a lot of commented code that can be used as examples for interacting
with the hardware. Credit to Niels here who got this all working.