Adds working Arduino Unit test suite for control package.
In this commit, we create a working test suite that tests some of the basic
functionality of the control package. Sadly, pretty much all of the tests
here are pretty arbitrary as they don't interact with the hardware. Sadly,
one significant limitation to these tests is they cannot be executed on a
native environment if there are any board specific libraries. This means that
in the future we will have to be isolating our arbitrary helper / utility methods
from files that contain direct hardware references if we want to test our code.