Added basic setup of Catch2 Testing Suite
The Catch2 testing suite is a very simple C++ unit testing suite that
requires the usage of a single header file `catch.hpp` in order to write
unit tests. This is a temporary endeavour until the team finalizes what
testing suite to use (PIO Unit Testing or something else). The usage of
Catch2 in this case is for demo purposes of gcov/gcovr only.