Updated .gitlab-ci.yml for test job
The important thing to note about this commit is the inclusion of job
artifacts. This means that the html files generated by gcovr (contains
code coverage info) is uploaded into GitLab, enabling us to download and
view the html files.

Also note that the job is isolated inside a `tests` folder. Test files
not in the tests folder will not be taken into account when the job is
executed. Furthermore, only files with `-tests` are acknowledged as a
test file by the script I wrote.