Installs SD libraries to lib_deps and renames cpp runner files
In this commit we install the SD libraries to lib_deps for the feather controller.
We also rename the *.{hpp,cpp} files introduced with the CPP code coverage runner
to .txt as they were causing platformio to fail to build. In later commit I will
attempt to add an exclusion rule for these