Includes sample code from STM32SD library in main.cpp
In this commit, we introduce example code from the stm32duino library
into the main.cpp file. To make room for this, we refactor the existing
dummy code for the MPU into its own respective method. The SD setup
code is introduced into a `void setup_sd(void)` method. The sample
code used within this commit can be seen at the following link:

https://github.com/stm32duino/STM32SD/blob/master/examples/ReadWrite/ReadWrite.ino