Creates Platformio subprojects
In this commit we move the contents from the first
project and move them into 2 different subprojects.
We have a 'development' and 'testing' project. The
development project is will contain all of the
finalised code whereas the testing project will
contain various methods to check whether the
hardware is performing correctly.