Integrates SD card tests into Platformio Unit Testing
In this commit we integrate Platformio Unit testing for the hardware level.
This turns out to be pretty finicky and hasn't been figured out fully yet.
When the tests begin executing, platformio will try upload the test files
to the board. It will then invoke the tests and they will run one by one.
However, the status of the tests is not displayed in the test window.

Instead the user is prompted with 'If you don’t see any output for the first
 10 secs, please reset board (press reset button)'. To get around this for
now we have to switch to the serial output perspective in platformio to
read the test results of the hardware. For now however, this is sufficient.
But it is definitely something that we should be looking into.