Updated code-coverage job and code coverage script
This commit updates the code-coverage CI/CD job to work with the new
version of the code coverage generation script. It also GitLab code
coverage parsing (see #121 for details) to enable GitLab to show % of
lines covered

This commit also sees a minor overhaul to the code coverage generation
script. The script now uses flags '-f' and '-t' to allow someone to
specify the directories of the files to test, and the directories to the
test files respectively (as opposed to relying solely on the position of
the argument when the script is executed to indicate whether or not its
a test target directory or a test directory). The script now supports an
arbitrary number of arguments, so long as the flags of the script are
used correctly. This allows us to obtain code coverage info on the
important directories of our PIO project (lib/hardware and lib/software)
all at once. I will update the README.md for code coverage generation in
the near-future.