Refactors IMU into new hardware libraries
In this commit, we refactor the IMU code into the new
libraries. We also fix a couple compiler warnings (would
have lead to memory leak) within the SD logic and tidy
up the main.cpp file. We also provide a simple example
of how it should be used including initialisation and
within loop().