Tests the code on the rocket hardware.
In this commit, we test the code on the hardware and
fix any of the issues that we were encountering. One
of the notable bugs that we had to fix was the Arduino
`printf` function does not support `%f`. Unfortunately,
due to issues with the SD card we are not able to verify
whether it is working as intended. The barometer is also
reading out slightly incorrect values, however the way
that we are interacting with it has not been changing.