Started v2 of uml state diagram for rocket
This commit creates a partially complete UML state diagram that revises
the state diagram currently located in section 4.1 of the architecture
document. The new version of the diagram aims to serve as a more
accurate version of the existing diagram by taking into account info we
now have since the last time the diagram was constructred.

Despite this, its important to note that its quite difficult to solidify
some of the states without further advice from the ECEN team. Also, some
of the technical details with regards to how mission control and the
rocket will communicate has not yet been firmly established at the time
of this commit. Particularly with how the software interlock state will
be exited by the rocket to begin its pre-launch setup.

Some of the ordering of the states may also need to be discussed with
the team (eg does the rocket begin its setup before or after receiving
the green light from mission control?). State transitions are highly
abstracted for the time being.