Renames test project directories, tidies code to meet project code style guide.
In this commit, we make a couple minor changes. Firstly we rename the testing
project folder names to match the rest of the project structure and naming
conventions. We also tidy the introduced main.cpp C++ code to touch up any
spelling mistakes and better comply with the Google Style Guide.