Implemented state pattern
This commit introduces some files to do with the state management of the
rocket. Currently, there are 2 important concepts, The StateMachine
class and the States class. The StateMachine construct serves the
purpose of executing the actions that need to be undertaken in a
specific state. State instances define a specific state the rocket can
take, and the actions the rocket must perform while in that state.

The advantage of implementing the state pattern to manage the state of
the rocket is that it enables the rest of the software system to issue
calls to the state machine to cause a change in state (eg in guidance
system, a call can be made to a StateMachine instance to force it to
enter the Landing state.) There is still some work to be done to ensure
the robustness of the state machine. For example, we need to ensure
that StateMachine can only be instantiated once throughout the life of
the rocket. We also need to enforce a sequential progression in states
as described by the UML state diagram of section 4.1 of the Architecture
Doc.

I've included some basic pseudocode in the implementation of the State
classes to give an idea of what kinds of things need to happen in each
state. Hopefully the rest of the team can give some feedback here and
add some of their own insight to better inform the software team on what
needs to happen in each state :)