Added test_environment_experiment PIO project
The purpose of this PIO project is to serve as a clean playground to
enable the exploration of PIO and its unit testing capabilities without
damaging the existing code base.

It is likely that this PIO project will serve as the proof-of-concept
with regards to our current unit testing for the entire project, as well
as the code coverage stuff we have setup so far. This is because of the
fact that this project is completely free from any adafruit libraries
which has proven to be an obstacle in the tackling of this issue.