Sets the Servo to a given angle via code.
In this commit, we add mildly tweaked sample code from the
library that can set both Servo's to a given angle.

For some stupid reason, I cannot seem get the `Servo` library
to work for me when installed as a project library. We got
around this for now by copying it across to the global library
directory: `~/.platformio/lib/...`.