Removed test_file_template.cpp and updated readme
Turns out putting the template somewhere safe for storage is not as
trivial as it seems. Placing it into the `code_coverage` directory meant
that the script was executing it as a regular test file, along with the
other test files. At this point, the simplest solution was to simply
place the template into the readme, which is what this commit does.