Implements mapping from upper to lower limits for servos.
In this commit, Billy and I play with the servos and find the
respective upper and lower limits for both of the servos. It is
worth noting that the limits differ for each of the servos, in
both offset and width. This means that we will need to keep
track of both. Currently we are writing to this using the
Arduino map function from 0 to 100 setting the value from a float
parsed via the serial input.