Adds mpu_servo_testing project to map mpu readings to servo.
In this commit, we add another test project where we map the
readings from the MPU (using the i2cdev and mpu-i2cdev) libraries.
Although we were having success with the green servo,
unfortunately the wiring has frayed on the pink servo causing
shorting meaning that we could not test it.