Integrates normal (non-pid) control system into development.
In this commit, we integrate a normal guidance system that
doesn't utilise a PID. Here we are using the same logic
that we have tested within one of our testing projects
`test_servo_mpu`. Here, we abstract some of the complexity
into classes following the structure of the rest of the
project.

Within this commit, we also aim fix some of the build
issues that had been experienced by some of the members
within the team. Part of this also included tidying up the
software and ensuring that it is in a tidy state to be
delivered. Apologies that this is such a big commit. To
implement these changes, I had created a new PIO project as
I was having lots of issues with PIO library management.